restructure of projects tests to organise into test type folders
