+++ .gitignore
@@ -18,6 +18,6 @@ node_modules/
 .dynamic/
 *ignore*
 # Ignore all PNGs in the snapshot folder
-/playwright-e2e/snapshots/tests/home-page.spec.ts-snapshots/*.png
+/playwright-e2e/snapshots/tests/**/*.png
 # But do NOT ignore PNGs that end in linux.png
-!/playwright-e2e/snapshots/tests/home-page.spec.ts-snapshots/*linux.png
+!/playwright-e2e/snapshots/tests/**/*linux.png
