+++ playwright-e2e/tests/functional-tests/book-a-recording/schedule-recording-functional.spec.ts
@@ -1,7 +1,7 @@
-import { test, expect } from '../../fixtures';
-import { config } from '../../utils';
+import { test, expect } from '../../../fixtures';
+import { config } from '../../../utils';
 
-test.describe('Set of tests to verify schedule a recording page for Level 1 user', () => {
+test.describe('Set of tests to verify functionality of schedule a recording page for Level 1 user', () => {
   const user = config.powerAppUsers.preLevel1User;
   test.use({ storageState: user.sessionFile });
