+++ playwright-e2e/tests/validation-tests/book-a-recording/case-details-validation.spec.ts
@@ -1,9 +1,9 @@
-import { test, expect } from '../../fixtures';
-import { BaseCaseDetails } from '../../types';
-import { config } from '../../utils';
+import { test, expect } from '../../../fixtures';
+import { BaseCaseDetails } from '../../../types';
+import { config } from '../../../utils';
 import { faker } from '@faker-js/faker';
 
-test.describe('Set of tests to verify case details page', () => {
+test.describe('Set of tests to verify validation of case details page is correct', () => {
   test.use({ storageState: config.powerAppUsers.preLevel1User.sessionFile });
 
   test.beforeEach(async ({ navigateToCaseDetailsPage }) => {
@@ -330,86 +330,6 @@ test.describe('Set of tests to verify case details page', () => {
         await expect(validationErrorText).toHaveText('Witness name must be between 1 and 25 characters.');
         await caseDetailsPage.$interactive.validationErrorCloseButton.click();
         await expect(validationErrorHeading).toBeHidden();
-      });
-    },
-  );
-
-  test(
-    'Verify when accessing the caseDetailsPage all the three buttons are in the correct state',
-
-    {
-      tag: '@Regression',
-    },
-
-    async ({ caseDetailsPage }) => {
-      await test.step('Verify all the buttons on case details page are visible', async () => {
-        await expect(caseDetailsPage.$interactive.modifyButton).toBeVisible();
-        await expect(caseDetailsPage.$interactive.saveButton).toBeVisible();
-        await expect(caseDetailsPage.$interactive.bookingsButton).toBeVisible();
-      });
-      await test.step('Verify modify and bookings buttons are disabled when accessing case details page', async () => {
-        await expect(caseDetailsPage.$interactive.modifyButton).toBeDisabled();
-        await expect(caseDetailsPage.$interactive.bookingsButton).toBeDisabled();
-      });
-      await test.step('Verify save button is enabled when accessing case details page', async () => {
-        await expect(caseDetailsPage.$interactive.saveButton).toBeEnabled();
-      });
-    },
-  );
-  test(
-    'Verify case details are correct when searching and selecting an exisiting case',
-    {
-      tag: '@Regression',
-    },
-    async ({ caseDetailsPage, apiClient }) => {
-      await test.step('Pre-requisite step in order to create a case via api', async () => {
-        await apiClient.createCase(2, 2);
-      });
-      const caseData = await apiClient.getCaseData();
-
-      await test.step('Verify case appears in searchlist when searched for', async () => {
-        await caseDetailsPage.$inputs.caseReference.fill(caseData.caseReference);
-        await expect(caseDetailsPage.$inputs.caseReference).toHaveValue(caseData.caseReference);
-
-        await expect(caseDetailsPage.$static.searchResultExistingCasesTitle).toBeVisible();
-        await expect(caseDetailsPage.$interactive.exsitingCaseFoundButtonInSearchList).toHaveCount(1);
-        await expect(caseDetailsPage.$interactive.exsitingCaseFoundButtonInSearchList).toContainText(caseData.caseReference);
-        await expect(caseDetailsPage.$interactive.exsitingCaseFoundButtonInSearchList).toContainText('Open');
-        await expect(caseDetailsPage.$interactive.exsitingCaseFoundButtonInSearchList).toContainText('PRE');
-        await expect(caseDetailsPage.$interactive.exsitingCaseFoundButtonInSearchList).toBeVisible();
-      });
-
-      await test.step('Verfiy test deatils are correct when exisiting case is selected from search list', async () => {
-        await caseDetailsPage.$interactive.exsitingCaseFoundButtonInSearchList.click();
-        await expect(caseDetailsPage.$static.selectedExistingCaseReferenceLable).toContainText(caseData.caseReference);
-        await expect(caseDetailsPage.$static.selectedExistingCaseReferenceLable).toBeVisible();
-
-        //Verify source lable contains PRE .
-        await expect(caseDetailsPage.$static.selectedExisitingCaseSourceLable).toContainText('PRE');
-        await expect(caseDetailsPage.$static.selectedExisitingCaseSourceLable).toBeVisible();
-
-        //Verify case status is Active.
-
-        await expect(caseDetailsPage.$static.selectedExisitingCaseStatusLable).toContainText('Active');
-        await expect(caseDetailsPage.$static.selectedExisitingCaseStatusLable).toBeVisible();
-
-        //Verify defendant value appears correctly.
-
-        const defendantValue = await caseDetailsPage.$inputs.defendants.inputValue();
-
-        for (const defendantName of caseData.defendantNames) {
-          expect(defendantValue).toContain(defendantName);
-        }
-        await expect(caseDetailsPage.$inputs.defendants).toBeVisible();
-
-        //Verify witness value appears correctly.
-
-        const witnessValue = await caseDetailsPage.$inputs.witnesses.inputValue();
-
-        for (const witnessName of caseData.witnessNames) {
-          expect(witnessValue).toContain(witnessName);
-        }
-        await expect(caseDetailsPage.$inputs.witnesses).toBeVisible();
       });
     },
   );
